refactor: put a token methods to a separate module
put a token methods to a separate module

MODULE-1-PART-3-gallery
